requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
blinker==1.6.2
cffi==1.15.1
click==8.1.7
comm==0.1.3
contourpy==1.0.6
cryptography==41.0.3
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.7
dj-database-url
Django==4.0.3
django-environ==0.11.2
executing==1.2.0
gunicorn==21.2.0
ipykernel
ipython
jedi==0.18.2
Pillow
psycopg2
djangorestframework==3.13.1
django-filter
whitenoise
